+++ src/lib/Scenes/Home/index.tsx
@@ -3,6 +3,8 @@ import { AppState, View } from "react-native"
 import ScrollableTabView from "react-native-scrollable-tab-view"
 import styled from "styled-components/native"
 
+import { Schema, screenTrack } from "lib/utils/track"
+
 import SwitchBoard from "lib/NativeModules/SwitchBoard"
 import { Router } from "lib/utils/router"
 
@@ -24,12 +26,25 @@ const TabBarContainer = styled.View`
 interface Props {
   selectedArtist?: string
   selectedTab?: number
+  tracking: any
 }
 
 interface State {
   appState: string
+  selectedTab: number
 }
 
+const ArtistsWorksForYouTab = 0
+const ForYouTab = 1
+const AuctionsTab = 2
+
+// This kills two birds with one stone:
+// It's necessary to wrap all tracks nested in this component, so they dispatch properly
+// Also, it'll only fire when the home screen is mounted, the only event we would otherwise miss with our own callbacks
+@screenTrack({
+  context_screen: Schema.PageNames.HomeArtistsWorksForYou,
+  context_screen_owner_type: null,
+})
 export default class Home extends React.Component<Props, State> {
   tabView?: ScrollableTabView | any
 
@@ -38,6 +53,7 @@ export default class Home extends React.Component<Props, State> {
 
     this.state = {
       appState: AppState.currentState,
+      selectedTab: ArtistsWorksForYouTab, // default to the WorksForYou view
     }
   }
 
@@ -49,20 +65,31 @@ export default class Home extends React.Component<Props, State> {
     AppState.removeEventListener("change", this._handleAppStateChange)
   }
 
+  // FIXME: Make a proper "viewDidAppear" callback / event emitter
+  // https://github.com/artsy/emission/issues/930
+  // This is called when the overall home component appears in Eigen
+  // We use it to dispatch screen events at that point
+  componentWillReceiveProps(newProps) {
+    if (newProps.isVisible) {
+      this.fireHomeScreenViewAnalytics()
+    }
+  }
+
   _handleAppStateChange = nextAppState => {
     // If we are coming back to the app from the background with a selected artist, make sure we're on the Artists tab
     if (this.props.selectedArtist && this.state.appState.match(/inactive|background/) && nextAppState === "active") {
-      this.tabView.goToPage(0)
+      this.tabView.goToPage(ArtistsWorksForYouTab)
     }
-    this.setState({ appState: nextAppState })
+    this.setState({ appState: nextAppState, selectedTab: ArtistsWorksForYouTab }, this.fireHomeScreenViewAnalytics)
   }
 
   render() {
     return (
       <View style={{ flex: 1 }}>
         <ScrollableTabView
-          initialPage={this.props.selectedTab || 0}
+          initialPage={this.props.selectedTab || ArtistsWorksForYouTab}
           ref={tabView => (this.tabView = tabView)}
+          onChangeTab={selectedTab => this.setSelectedTab(selectedTab)}
           renderTabBar={props => (
             <TabBarContainer>
               <TabBar {...props} />
@@ -96,5 +123,25 @@ export default class Home extends React.Component<Props, State> {
 
   openLink() {
     SwitchBoard.presentNavigationViewController(this, Router.ConsignmentsStartSubmission)
+  }
+
+  setSelectedTab(selectedTab) {
+    this.setState({ selectedTab: selectedTab.i }, this.fireHomeScreenViewAnalytics)
+  }
+
+  fireHomeScreenViewAnalytics = () => {
+    let screenType
+
+    if (this.state.selectedTab === ArtistsWorksForYouTab) {
+      screenType = Schema.PageNames.HomeArtistsWorksForYou
+    } else if (this.state.selectedTab === ForYouTab) {
+      screenType = Schema.PageNames.HomeForYou
+    } else if (this.state.selectedTab === AuctionsTab) {
+      screenType = Schema.PageNames.HomeAuctions
+    }
+    this.props.tracking.trackEvent({
+      context_screen: screenType,
+      context_screen_owner_type: null,
+    })
   }
 }