+++ CHANGELOG.md
@@ -6,9 +6,11 @@
 * Make sure Home/Fairs rail does not clip any icons at the end of the list - alloy
 * Re-align headers in all Home tabs - alloy
 * Fix crash that would occur when tapping artwork component on inquiry view - alloy
+* Add analytics to the Home tabs - maxim
 * Fix a bug that would not let an infinite scroll of artworks stop requesting more artworks - orta
 * More pixel pushing in the consignments welcome screen - alloy
 * More pixel pushing in the conversations zero state screen - alloy
+* Add saves and follows analytics - maxim
 
 ### 1.4.0-beta.13